Refactor resource_manager to have a common one for all architectures
